requirements.txt
yara-python==4.5.0
watchdog
psutil
requests
setuptools
build
